added type hints, changed filenames, deleted sample, len, get_item
