+++ tests/test_diffeq/test_wrappedscipysolver_cases/test_wrappedscipysolver_cases.py
@@ -3,14 +3,14 @@ import pytest_cases
 from scipy.integrate._ivp import rk
 
 from probnum import diffeq
-from probnum.diffeq import wrapperscipysolver
+from probnum.diffeq import wrappedscipysolver
 
 
 def case_lorenz():
     y0 = np.array([0.0, 1.0, 1.05])
     ode = diffeq.lorenz([0.0, 1.0], y0)
     scipysolver = rk.RK45(ode.rhs, ode.t0, y0, ode.tmax)
-    testsolver = wrapperscipysolver.WrapperScipyRungeKutta(
+    testsolver = wrappedscipysolver.WrappedScipyRungeKutta(
         rk.RK45(ode.rhs, ode.t0, y0, ode.tmax), order=4
     )
     return testsolver, scipysolver
@@ -20,7 +20,7 @@ def case_logistic():
     y0 = np.array([0.1])
     ode = diffeq.logistic([0.0, 1.0], y0)
     scipysolver = rk.RK45(ode.rhs, ode.t0, y0, ode.tmax)
-    testsolver = wrapperscipysolver.WrapperScipyRungeKutta(
+    testsolver = wrappedscipysolver.WrappedScipyRungeKutta(
         rk.RK45(ode.rhs, ode.t0, y0, ode.tmax), order=4
     )
     return testsolver, scipysolver
@@ -30,7 +30,7 @@ def case_lotkavolterra():
     y0 = np.array([0.1, 0.1])
     ode = diffeq.lotkavolterra([0.0, 1.0], y0)
     scipysolver = rk.RK45(ode.rhs, ode.t0, y0, ode.tmax)
-    testsolver = wrapperscipysolver.WrapperScipyRungeKutta(
+    testsolver = wrappedscipysolver.WrappedScipyRungeKutta(
         rk.RK45(ode.rhs, ode.t0, y0, ode.tmax), order=4
     )
     return testsolver, scipysolver
